Initial commit of prototype node project and python importer
Node project basically only looks at the stripe api in an attempt to pull in information.. however it's become apparant that the interaction between Squarespace and Stripe is reference only.

So, I've created a python script to take raw csv orders and ready them (for now) in a dictionary. Heavy lifting is done by the built in csv lib in python.