Initial commit of membership intake form progress
Form isn't done yet but this is what I have. We'll use an updates folder and date stamped files to keep track of db changes, we'll use the databse-controller.js module to drive database interaction, and the dist folder to manage files destined for remote invocation. The testing folder will hold unit tests, and html files mimicing a ss environment.