Put more work into the remote order intake form
I keep hitting walls with Squarespace's support for custom code. Now it seems that the checkout page is off limits. Every page up to the checkout page is fine... but not the checkout page.

I'm starting to wonder if this will be worth the work. There are many considerations to take when modifying the add to cart flow (cart abandondment, removing items, custom checkout forms). We may be closer to running our own portal now....